+++ examples/main/shared/editor/Editor.tsx
@@ -19,6 +19,7 @@ import {
 } from 'wix-rich-content-common';
 import { TestAppConfig } from '../../src/types';
 import { RicosEditor, RicosEditorProps, RicosEditorType } from 'ricos-editor';
+import createSideBlockComponent from '../../src/Components/createSideBlockComponent';
 
 const STATIC_TOOLBAR = 'static';
 
@@ -39,6 +40,7 @@ interface ExampleEditorProps {
   experiments?: AvailableExperiments;
   externalPopups: boolean;
   textWrap?: boolean;
+  showSideBlockComponent?: boolean;
 }
 
 export default class Editor extends PureComponent<ExampleEditorProps> {
@@ -156,6 +158,7 @@ export default class Editor extends PureComponent<ExampleEditorProps> {
       experiments,
       externalPopups,
       textWrap,
+      showSideBlockComponent,
     } = this.props;
     const textToolbarType: TextToolbarType = staticToolbar && !isMobile ? STATIC_TOOLBAR : null;
     const useStaticTextToolbar = textToolbarType === STATIC_TOOLBAR;
@@ -185,6 +188,9 @@ export default class Editor extends PureComponent<ExampleEditorProps> {
             experiments={experiments}
             textWrap={textWrap}
             onAtomicBlockFocus={d => console.log('onAtomicBlockFocus', d)} // eslint-disable-line
+            sideBlockComponent={
+              showSideBlockComponent && createSideBlockComponent(this.editor?.getEditorCommands())
+            }
           />
         </div>
       </div>