+++ examples/storybook/src/shared/resources/experimentsExample.ts
@@ -1,9 +1,9 @@
 export const EXPERIMENTS = [
   { name: 'newFormattingToolbar', scope: 'ricos' },
   { name: 'tiptapEditor', scope: 'ricos' },
-  { name: 'newVideoModal', scope: 'ricos' },
-  { name: 'newVerticalEmbedModal', scope: 'ricos' },
-  { name: 'newSocialEmbedModal', scope: 'ricos' },
+  { name: 'fixListLineHeight', scope: 'ricos' },
+  { name: 'newVideoVerticalAndSocialModals', scope: 'ricos' },
+  { name: 'fixedTabSize', scope: 'ricos' },
   { name: 'forceIframeSandboxDomain', scope: 'ricos' },
   { name: 'useSeo', scope: 'wixRicos' },
   { name: 'spoilerInSettings', scope: 'ricos' },