+++ packages/editor/web/src/RichContentEditor/RichContentEditor.tsx
@@ -201,15 +201,6 @@ interface RichContentEditorState {
   undoRedoStackChanged: boolean;
 }
 
-// experiment example code
-function makeBarrelRoll() {
-  document.querySelector('.DraftEditor-root')?.classList.toggle('barrelRoll');
-  setTimeout(
-    () => document.querySelector('.DraftEditor-root')?.classList.toggle('barrelRoll'),
-    1000
-  );
-}
-
 class RichContentEditor extends Component<RichContentEditorProps, RichContentEditorState> {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   refId: number;
@@ -918,13 +909,6 @@ class RichContentEditor extends Component<RichContentEditorProps, RichContentEdi
       modifiers: [MODIFIERS.COMMAND, MODIFIERS.SHIFT],
       key: 'z',
     },
-    this.props.experiments?.barrelRoll?.enabled && typeof window !== 'undefined'
-      ? {
-          command: 'cmdShift7',
-          modifiers: [MODIFIERS.COMMAND, MODIFIERS.SHIFT],
-          key: '7',
-        }
-      : {},
   ] as KeyCommand[];
 
   customCommandHandlers: Record<string, CommandHandler> = {
@@ -935,9 +919,6 @@ class RichContentEditor extends Component<RichContentEditorProps, RichContentEdi
     esc: this.handleEscCommand,
     ricosUndo: this.handleUndoCommand,
     ricosRedo: this.handleRedoCommand,
-    ...(this.props.experiments?.barrelRoll?.enabled && typeof window !== 'undefined'
-      ? { cmdShift7: makeBarrelRoll }
-      : {}),
   };
 
   getCustomCommandHandlers = (): {
@@ -1135,6 +1116,7 @@ class RichContentEditor extends Component<RichContentEditorProps, RichContentEdi
       handleReturn,
       readOnly,
       onBackspace,
+      experiments,
     } = this.props;
     const { editorState } = this.state;
     const { theme } = this.contextualData;
@@ -1152,7 +1134,12 @@ class RichContentEditor extends Component<RichContentEditorProps, RichContentEdi
         handleBeforeInput={this.handleBeforeInput}
         handlePastedText={this.handlePastedText}
         plugins={this.plugins}
-        blockStyleFn={blockStyleFn(theme, this.styleToClass, textAlignment)}
+        blockStyleFn={blockStyleFn(
+          theme,
+          this.styleToClass,
+          textAlignment,
+          experiments?.fixedTabSize?.enabled
+        )}
         handleKeyCommand={handleKeyCommand(
           this.updateEditorState,
           this.getCustomCommandHandlers().commandHandlers,
@@ -1340,6 +1327,7 @@ class RichContentEditor extends Component<RichContentEditorProps, RichContentEdi
       showToolbars = true,
       isInnerRCE,
       isMobile = false,
+      experiments,
     } = this.props;
     const { innerModal } = this.state;
 
@@ -1362,6 +1350,7 @@ class RichContentEditor extends Component<RichContentEditorProps, RichContentEdi
         theme,
         draftStyles,
         editorStyles,
+        experiments,
       });
 
       return (