+++ packages/plugin-image/web/src/tiptap/tiptap.ts
@@ -33,6 +33,9 @@ export const createRicosExtensions: CreateRicosExtensions = defaultOptions => [
     createExtensionConfig: () => ({
       name,
       atom: false,
+      group: 'block',
+      selectable: true,
+      draggable: true,
       addOptions: () => defaultOptions,
       addCommands() {
         return {