+++ packages/tiptap-editor/web/src/models/NodeExtension.tsx
@@ -1,5 +1,8 @@
 import {
   mergeAttributes,
+  textblockTypeInputRule,
+  markPasteRule,
+  markInputRule,
   Node,
   NodeConfig,
   NodeViewRendererProps,
@@ -18,14 +21,15 @@ import { isRicosNodeExtension, RicosExtension, RicosNodeExtension } from 'ricos-
 import { RicosNode } from '../components/RicosNode';
 
 const toExtensionConfig = (ext: RicosNodeExtension) =>
-  ext.createExtensionConfig({ mergeAttributes });
+  ext.createExtensionConfig({
+    textblockTypeInputRule,
+    markPasteRule,
+    markInputRule,
+    mergeAttributes,
+  });
 
 const toFullNodeConfig = (ext: RicosNodeExtension) => (config: NodeConfig): NodeConfig => ({
   type: 'node',
-  group: 'block',
-  atom: true,
-  selectable: true,
-  draggable: true,
   parseHTML: () => [{ tag: `${config.name}-component` }],
   renderHTML: ({ HTMLAttributes }) => [`${config.name}-component`, mergeAttributes(HTMLAttributes)],
   addAttributes: () => ext.componentDataDefaults || {},