+++ packages/viewer/web/src/RichContentViewer.tsx
@@ -165,6 +165,7 @@ class RichContentViewer extends Component<
       seoMode,
       iframeSandboxDomain,
       textAlignment,
+      experiments,
     }: RichContentViewerProps,
     contentState?: DraftContent
   ): ViewerContextType => {
@@ -192,6 +193,7 @@ class RichContentViewer extends Component<
       iframeSandboxDomain,
       disableRightClick: config?.uiSettings?.disableRightClick,
       textAlignment,
+      experiments,
     };
   };
 
@@ -251,6 +253,7 @@ class RichContentViewer extends Component<
       typeMappers,
       setRef = () => {},
       onMouseOver = () => {},
+      experiments = {},
     } = this.props;
     const decorators = [...this.props.decorators, createJustificationFixDecorator()];
     try {
@@ -267,8 +270,10 @@ class RichContentViewer extends Component<
         viewerStyles.renderedInTable,
         draftDefaultStyles.renderedInTable
       );
+
       const editorClassName = classNames(styles.editor, renderedInTable && tableClassNames, {
         [styles.rtl]: textDirection === 'rtl',
+        [styles.fixedTabSize]: experiments.fixedTabSize?.enabled,
       });
 
       const initSpoilers = config[SPOILER_TYPE]?.initSpoilersContentState;
@@ -294,7 +299,7 @@ class RichContentViewer extends Component<
         config,
         initSpoilers,
         SpoilerViewerWrapper,
-        { addAnchors },
+        { addAnchors, fixedTabSize: experiments.fixedTabSize?.enabled },
         innerRCEViewerProps,
         this.props.documentStyle
       );