+++ packages/viewer/web/src/utils/convertContentState.tsx
@@ -61,7 +61,14 @@ const getInline = (contentState, config, inlineStyleMappers, mergedStyles) =>
     mergedStyles
   );
 
-const getBlocks = (mergedStyles, textDirection, context, addAnchorsPrefix, documentStyle) => {
+const getBlocks = (
+  mergedStyles,
+  textDirection,
+  context,
+  addAnchorsPrefix,
+  documentStyle,
+  fixedTabSize
+) => {
   const getList = ordered => (items, blockProps) => {
     const fixedItems = items.map(item => (item.length ? item : [' ']));
 
@@ -80,6 +87,7 @@ const getBlocks = (mergedStyles, textDirection, context, addAnchorsPrefix, docum
         };
       },
       context,
+      fixedTabSize,
     };
     return <List {...props} />;
   };
@@ -128,6 +136,7 @@ const getBlocks = (mergedStyles, textDirection, context, addAnchorsPrefix, docum
                   mergedStyles[style]
                 ),
                 depthClassName(depth),
+                { 'fixed-tab-size': fixedTabSize },
                 directionBlockClassName,
                 isPaywallSeo(context.seoMode) &&
                   getPaywallSeoClass(context.seoMode.paywall, blockIndex)
@@ -277,7 +286,7 @@ const convertToReact = (
   config: LegacyViewerPluginConfig,
   initSpoilers: (content?: DraftContent) => DraftContent | undefined,
   SpoilerViewerWrapper: unknown,
-  options: { addAnchors?: boolean | string; [key: string]: unknown } = {},
+  options: { addAnchors?: boolean | string; [key: string]: unknown; fixedTabSize?: boolean } = {},
   innerRCEViewerProps?: {
     typeMappers: PluginTypeMapper[];
     inlineStyleMappers: InlineStyleMapperFunction[];
@@ -290,7 +299,7 @@ const convertToReact = (
   if (isEmptyContentState(context.contentState)) {
     return null;
   }
-  const { addAnchors, ...restOptions } = options;
+  const { addAnchors, fixedTabSize, ...restOptions } = options;
   const normalizedContentState = context.contentState
     ? normalizeContentState(context.contentState)
     : context.contentState;
@@ -305,7 +314,14 @@ const convertToReact = (
     newContentState,
     {
       inline: getInline(newContentState, config, inlineStyleMappers, mergedStyles),
-      blocks: getBlocks(mergedStyles, textDirection, context, addAnchorsPrefix, documentStyle),
+      blocks: getBlocks(
+        mergedStyles,
+        textDirection,
+        context,
+        addAnchorsPrefix,
+        documentStyle,
+        fixedTabSize
+      ),
       entities: getEntities(
         typeMappers,
         context,